// src/types.ts
export type MuscleGroup =
  | 'chest' | 'shoulders' | 'triceps'
  | 'biceps' | 'forearms'
  | 'lats' | 'traps' | 'lower_back'
  | 'abs' | 'obliques'
  | 'quads' | 'hamstrings' | 'glutes' | 'calves';